requirements/flask-adapters.txt
-f dist
finitelycomputable-helloworld-flask[tests]~=24.3
finitelycomputable-helloworld-cherrypy-flask[tests]~=24.3
finitelycomputable-helloworld-falcon-flask[tests]~=24.3
finitelycomputable-helloworld-morepath-flask[tests]~=24.3
finitelycomputable-helloworld-quart-flask[tests]~=24.3

finitelycomputable-idtrust-flask-peewee[tests]~=24.3

finitelycomputable-cherrypy-mount[tests]~=24.3
finitelycomputable-falcon-addroute[tests]~=24.3
finitelycomputable-flask-blueprints[tests]~=24.3
finitelycomputable-flask-dispatcher[tests]~=24.3
finitelycomputable-morepath-mount[tests]~=24.3
